+++ Kursova_VideoStore/Controllers/OrdersController.cs
@@ -4,7 +4,6 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Videoteka.Models;
@@ -22,29 +21,76 @@ namespace Kursova_VideoStore.Controllers
         }
 
         // GET: Orders
-        public async Task<IActionResult> Index()
+        public async Task<IActionResult> Index(
+            string sortOrder,
+            string currentFilter,
+            string searchString,
+            int? pageNumber)
         {
-            var videotekaContext = _context.Orders.Include(o => o.Customer).Include(o => o.Employee);
-            return View(await videotekaContext.ToListAsync());
+            ViewData["CurrentSort"] = sortOrder;
+            ViewData["DateSortParm"] = String.IsNullOrEmpty(sortOrder) ? "date_desc" : "";
+            ViewData["CustomerSortParm"] = sortOrder == "Customer" ? "customer_desc" : "Customer";
+            ViewData["EmployeeSortParm"] = sortOrder == "Employee" ? "employee_desc" : "Employee";
+
+            if (searchString != null)
+            {
+                pageNumber = 1;
+            }
+            else
+            {
+                searchString = currentFilter;
+            }
+
+            ViewData["CurrentFilter"] = searchString;
+
+            // Include related data (Customer and Employee)
+            var orders = _context.Orders
+                .Include(o => o.Customer)
+                .Include(o => o.Employee)
+                .AsQueryable();
+
+            if (!String.IsNullOrEmpty(searchString))
+            {
+                orders = orders.Where(s => s.Customer.LastName.Contains(searchString)
+                                       || s.Customer.FirstName.Contains(searchString)
+                                       || s.Employee.LastName.Contains(searchString));
+            }
+
+            switch (sortOrder)
+            {
+                case "date_desc":
+                    orders = orders.OrderByDescending(s => s.OrderDate);
+                    break;
+                case "Customer":
+                    orders = orders.OrderBy(s => s.Customer.LastName);
+                    break;
+                case "customer_desc":
+                    orders = orders.OrderByDescending(s => s.Customer.LastName);
+                    break;
+                case "Employee":
+                    orders = orders.OrderBy(s => s.Employee.LastName);
+                    break;
+                case "employee_desc":
+                    orders = orders.OrderByDescending(s => s.Employee.LastName);
+                    break;
+                default:
+                    orders = orders.OrderBy(s => s.OrderDate);
+                    break;
+            }
+
+            int pageSize = 5;
+            return View(await PaginatedList<Order>.CreateAsync(orders.AsNoTracking(), pageNumber ?? 1, pageSize));
         }
 
         // GET: Orders/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
+            if (id == null) return NotFound();
             var order = await _context.Orders
                 .Include(o => o.Customer)
                 .Include(o => o.Employee)
                 .FirstOrDefaultAsync(m => m.OrderID == id);
-            if (order == null)
-            {
-                return NotFound();
-            }
-
+            if (order == null) return NotFound();
             return View(order);
         }
 
@@ -52,17 +98,16 @@ namespace Kursova_VideoStore.Controllers
         public IActionResult Create()
         {
             ViewData["CustomerID"] = new SelectList(_context.Customers, "CustomerID", "Email");
-            ViewData["EmployeeID"] = new SelectList(_context.Employees, "EmployeeID", "EmployeeID");
+            ViewData["EmployeeID"] = new SelectList(_context.Employees, "EmployeeID", "FirstName");
             return View();
         }
 
         // POST: Orders/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("OrderID,CustomerID,EmployeeID,OrderDate")] Order order)
         {
+            // Note: In real scenarios, you might auto-set OrderDate here
             if (ModelState.IsValid)
             {
                 _context.Add(order);
@@ -70,39 +115,27 @@ namespace Kursova_VideoStore.Controllers
                 return RedirectToAction(nameof(Index));
             }
             ViewData["CustomerID"] = new SelectList(_context.Customers, "CustomerID", "Email", order.CustomerID);
-            ViewData["EmployeeID"] = new SelectList(_context.Employees, "EmployeeID", "EmployeeID", order.EmployeeID);
+            ViewData["EmployeeID"] = new SelectList(_context.Employees, "EmployeeID", "FirstName", order.EmployeeID);
             return View(order);
         }
 
         // GET: Orders/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
+            if (id == null) return NotFound();
             var order = await _context.Orders.FindAsync(id);
-            if (order == null)
-            {
-                return NotFound();
-            }
+            if (order == null) return NotFound();
             ViewData["CustomerID"] = new SelectList(_context.Customers, "CustomerID", "Email", order.CustomerID);
-            ViewData["EmployeeID"] = new SelectList(_context.Employees, "EmployeeID", "EmployeeID", order.EmployeeID);
+            ViewData["EmployeeID"] = new SelectList(_context.Employees, "EmployeeID", "FirstName", order.EmployeeID);
             return View(order);
         }
 
         // POST: Orders/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("OrderID,CustomerID,EmployeeID,OrderDate")] Order order)
         {
-            if (id != order.OrderID)
-            {
-                return NotFound();
-            }
+            if (id != order.OrderID) return NotFound();
 
             if (ModelState.IsValid)
             {
@@ -113,39 +146,25 @@ namespace Kursova_VideoStore.Controllers
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!OrderExists(order.OrderID))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    if (!OrderExists(order.OrderID)) return NotFound();
+                    else throw;
                 }
                 return RedirectToAction(nameof(Index));
             }
             ViewData["CustomerID"] = new SelectList(_context.Customers, "CustomerID", "Email", order.CustomerID);
-            ViewData["EmployeeID"] = new SelectList(_context.Employees, "EmployeeID", "EmployeeID", order.EmployeeID);
+            ViewData["EmployeeID"] = new SelectList(_context.Employees, "EmployeeID", "FirstName", order.EmployeeID);
             return View(order);
         }
 
         // GET: Orders/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
+            if (id == null) return NotFound();
             var order = await _context.Orders
                 .Include(o => o.Customer)
                 .Include(o => o.Employee)
                 .FirstOrDefaultAsync(m => m.OrderID == id);
-            if (order == null)
-            {
-                return NotFound();
-            }
-
+            if (order == null) return NotFound();
             return View(order);
         }
 
@@ -155,11 +174,7 @@ namespace Kursova_VideoStore.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var order = await _context.Orders.FindAsync(id);
-            if (order != null)
-            {
-                _context.Orders.Remove(order);
-            }
-
+            if (order != null) _context.Orders.Remove(order);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
